- renamed files codonsimulator.* to codon_simulator.*
