+++ src/CMakeLists.txt
@@ -10,7 +10,7 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++17 -fPIC")
 
 
 set(common_sources concentrationsreader.cpp mrna_reader.cpp)
-set(codon_simulator_sources codonsimulator.cpp ${common_sources})
+set(codon_simulator_sources codon_simulator.cpp ${common_sources})
 set(elongation_simulator_sources   elongation_codon.cpp initiationterminationcodon.cpp mrnaelement.cpp sequence_simulator.cpp ${codon_simulator_sources} elongation_simulation_manager.cpp elongation_simulation_processor.cpp)
 add_library(util_lib STATIC ${common_sources})
 add_library(codon_simulator_lib STATIC ${codon_simulator_sources})
@@ -37,23 +37,23 @@ if(BUILD_JULIA)
 endif(BUILD_JULIA)
 
 add_library(sequence_simulator EXCLUDE_FROM_ALL MODULE sequence_simulator.cpp elongation_simulation_manager.cpp elongation_simulation_processor.cpp mrna_reader.cpp initiationterminationcodon.cpp
-concentrationsreader.cpp codonsimulator.cpp mrnaelement.cpp elongation_codon.cpp)
+concentrationsreader.cpp codon_simulator.cpp mrnaelement.cpp elongation_codon.cpp)
 target_compile_definitions(sequence_simulator PRIVATE -DCOMIPLE_PYTHON_MODULE) # enable python pathway in code.
 target_link_libraries(sequence_simulator PRIVATE pybind11::module ${PYTHON_LIBRARIES} jsoncpp)
 set_target_properties(sequence_simulator PROPERTIES PREFIX "${PYTHON_MODULE_PREFIX}"
                                           SUFFIX "${PYTHON_MODULE_EXTENSION}")
 
-add_library(codonsimulator EXCLUDE_FROM_ALL MODULE concentrationsreader.cpp codonsimulator.cpp)
+add_library(codon_simulator EXCLUDE_FROM_ALL MODULE concentrationsreader.cpp codon_simulator.cpp)
 
-target_compile_definitions(codonsimulator PRIVATE -DCOMIPLE_PYTHON_MODULE ) # enable python pathway in code.
+target_compile_definitions(codon_simulator PRIVATE -DCOMIPLE_PYTHON_MODULE ) # enable python pathway in code.
 # Julia interface
 if(BUILD_JULIA)
-    target_compile_definitions(codonsimulator PRIVATE -DCOMIPLE_JULIA_MODULE ) # enable Julia pathway in code.
-    target_link_libraries(codonsimulator PRIVATE JlCxx::cxxwrap_julia)
+    target_compile_definitions(codon_simulator PRIVATE -DCOMIPLE_JULIA_MODULE ) # enable Julia pathway in code.
+    target_link_libraries(codon_simulator PRIVATE JlCxx::cxxwrap_julia)
 endif(BUILD_JULIA)
 
-target_link_libraries(codonsimulator PRIVATE pybind11::module ${PYTHON_LIBRARIES})
-set_target_properties(codonsimulator PROPERTIES PREFIX "${PYTHON_MODULE_PREFIX}"
+target_link_libraries(codon_simulator PRIVATE pybind11::module ${PYTHON_LIBRARIES})
+set_target_properties(codon_simulator PROPERTIES PREFIX "${PYTHON_MODULE_PREFIX}"
                                           SUFFIX "${PYTHON_MODULE_EXTENSION}")
